refactor: extracted time stamp logic into separate util
