+++ client/src/utils/userManagement.ts
@@ -3,6 +3,7 @@ import { AddUserMutation } from '../gql/mutations/addUsers'
 import { GetUserListQuery } from '../gql/queries/userList'
 import { DeleteUserMutation } from '../gql/mutations/deleteUsers'
 import { UnenrolUsersFromQuizMutation } from '../gql/mutations/userQuiz'
+import { generateTimestamp, generateDownloadLink } from './time_stamp'
 
 export const addUserMutation = async (
   apollo: ApolloClient<NormalizedCacheObject>,
@@ -111,27 +112,10 @@ export const downloadUsersCsvQuery = async (
     // Create a temporary link element
     const link = document.createElement('a')
     link.href = URL.createObjectURL(blob)
-    const currentDate = new Date()
-    const year = currentDate.getFullYear().toString().padStart(4, '0')
-    const month = (currentDate.getMonth() + 1).toString().padStart(2, '0')
-    const day = currentDate.getDate().toString().padStart(2, '0')
-    const hour = currentDate.getHours().toString().padStart(2, '0')
-    const minute = currentDate.getMinutes().toString().padStart(2, '0')
-    const second = currentDate.getSeconds().toString().padStart(2, '0')
-    link.download =
-      'all-students_' +
-      year +
-      '-' +
-      month +
-      '-' +
-      day +
-      '-' +
-      hour +
-      '-' +
-      minute +
-      '-' +
-      second +
-      '.csv'
+
+    // Specify name of downloaded csv
+    const timestamp = generateTimestamp()
+    link.download = generateDownloadLink('all-students', timestamp, 'csv')
 
     // Programmatically click the link to trigger the download
     link.click()
